jwt authorization added

+++ src/controllers/box.controller.ts
@@ -1,31 +1,30 @@
-import {
-  Count,
-  CountSchema,
-  Filter,
-  FilterExcludingWhere,
-  repository,
-  Where,
-} from '@loopback/repository';
+import {authenticate} from '@loopback/authentication';
+import {inject} from '@loopback/core';
+import {Filter, FilterExcludingWhere, repository} from '@loopback/repository';
 import {
   del,
   get,
   getModelSchemaRef,
+  HttpErrors,
   param,
   patch,
   post,
-  put,
   requestBody,
   response,
 } from '@loopback/rest';
+import {SecurityBindings, securityId, UserProfile} from '@loopback/security';
 import {Box} from '../models';
-import {BoxRepository} from '../repositories';
+import {BoxRepository, UserRepository} from '../repositories';
 
 export class BoxController {
   constructor(
     @repository(BoxRepository)
     public boxRepository: BoxRepository,
+    @repository(UserRepository)
+    public userRepository: UserRepository,
   ) {}
 
+  @authenticate('jwt')
   @post('/boxes')
   @response(200, {
     description: 'Box model instance',
@@ -43,22 +42,19 @@ export class BoxController {
       },
     })
     box: Omit<Box, 'id'>,
+    @inject(SecurityBindings.USER) currentUserProfile: UserProfile,
   ): Promise<Box> {
+    const userId = currentUserProfile[securityId];
+    const user = await this.userRepository.findById(userId);
+    if (user.login !== box.sender) {
+      throw new HttpErrors.Forbidden('Access denied');
+    }
     return await this.boxRepository.create({
       ...box,
       status: 'not deployed',
       createdAt: Date.now(),
       updatedAt: Date.now(),
     });
-  }
-
-  @get('/boxes/count')
-  @response(200, {
-    description: 'Box model count',
-    content: {'application/json': {schema: CountSchema}},
-  })
-  async count(@param.where(Box) where?: Where<Box>): Promise<Count> {
-    return this.boxRepository.count(where);
   }
 
   @get('/boxes')
@@ -74,28 +70,10 @@ export class BoxController {
     },
   })
   async find(@param.filter(Box) filter?: Filter<Box>): Promise<Box[]> {
-    return this.boxRepository.find(filter);
+    return this.boxRepository.find({...filter, fields: {secret: false}});
   }
 
-  @patch('/boxes')
-  @response(200, {
-    description: 'Box PATCH success count',
-    content: {'application/json': {schema: CountSchema}},
-  })
-  async updateAll(
-    @requestBody({
-      content: {
-        'application/json': {
-          schema: getModelSchemaRef(Box, {partial: true}),
-        },
-      },
-    })
-    box: Box,
-    @param.where(Box) where?: Where<Box>,
-  ): Promise<Count> {
-    return this.boxRepository.updateAll(box, where);
-  }
-
+  @authenticate('jwt')
   @get('/boxes/{id}')
   @response(200, {
     description: 'Box model instance',
@@ -107,11 +85,22 @@ export class BoxController {
   })
   async findById(
     @param.path.string('id') id: string,
+    @inject(SecurityBindings.USER) currentUserProfile: UserProfile,
     @param.filter(Box, {exclude: 'where'}) filter?: FilterExcludingWhere<Box>,
   ): Promise<Box> {
+    const userId = currentUserProfile[securityId];
+    const user = await this.userRepository.findById(userId);
+    const box = await this.boxRepository.findById(id);
+    if (user.login !== box.sender && user.login !== box.reciever) {
+      throw new HttpErrors.Forbidden('Access denied');
+    }
+    if (box.status === 'first deployed' || box.status === 'not deployed') {
+      filter = {...filter, fields: {secret: false}};
+    }
     return this.boxRepository.findById(id, filter);
   }
 
+  @authenticate('jwt')
   @patch('/boxes/{id}')
   @response(204, {
     description: 'Box PATCH success',
@@ -126,26 +115,35 @@ export class BoxController {
       },
     })
     box: Box,
+    @inject(SecurityBindings.USER) currentUserProfile: UserProfile,
   ): Promise<void> {
+    const userId = currentUserProfile[securityId];
+    const user = await this.userRepository.findById(userId);
+    const oldBox = await this.boxRepository.findById(id);
+    if (user.login === oldBox.sender && user.login === box.reciever) {
+      throw new HttpErrors.Conflict('Reciever and sender cannot be equal');
+    }
     await this.boxRepository.updateById(id, box);
   }
 
-  @put('/boxes/{id}')
-  @response(204, {
-    description: 'Box PUT success',
-  })
-  async replaceById(
-    @param.path.string('id') id: string,
-    @requestBody() box: Box,
-  ): Promise<void> {
-    await this.boxRepository.replaceById(id, box);
-  }
-
+  @authenticate('jwt')
   @del('/boxes/{id}')
   @response(204, {
     description: 'Box DELETE success',
   })
-  async deleteById(@param.path.string('id') id: string): Promise<void> {
+  async deleteById(
+    @param.path.string('id') id: string,
+    @inject(SecurityBindings.USER) currentUserProfile: UserProfile,
+  ): Promise<void> {
+    const userId = currentUserProfile[securityId];
+    const user = await this.userRepository.findById(userId);
+    const box = await this.boxRepository.findById(id);
+    if (user.login !== box.sender) {
+      throw new HttpErrors.Forbidden('Access denied');
+    }
+    if (box.status !== 'not deployed') {
+      throw new HttpErrors.Conflict('You cannot delete box after deploy');
+    }
     await this.boxRepository.deleteById(id);
   }
 }
